Working on the new indexes. Added some tests.
I need to work on providing the option to generate migrations of some 
kind, therefore allowing the system to automatically upgrade 
installations of a certain software without the need for external help.

The thing is that, in order to generate migrations we would need one of 
the following:

A: Reverse engineer the current state of the DB. Create a schema, 
compare to the provided one and write a migration.

B: Let the user "commit" the current schema. Allowing them to rollback
to a previous state if needed.

Currently B sounds like a way better option in my mind, and I'll be 
looking forward to implement it once the whole thing is more advanced.
I feel like SF should have a way better CLI interface at this point.

-- I'm personally rather happy with how this is looking so far. Breaking
the database abstraction into many smaller pieces is making them more 
manageable and less prone to breaking. That's, IMO, a big step in the 
right direction. Specially because database stuff is very delicate.