Imported my first little experiments with the validation compiler. The idea here is to provide a simple yet powerful way of validating IO from a static source like doc-blocks are.
Usually, writing validation code is a messy task. Even when you set up a
"smart" kind of API for it, the code tends to look like a huge messy
and unmantainable blob.

I'm looking forward to have validation executed before the controller is
invoked. Making it easier and more convenient to maintain.

-- Currently my concern is, that in the event of the docblock parser
failing, or any other technical issue, the validation would not get
executed and the application would potentially be unaware of the issue.
Leaving the application "vulnerable" to invalid user data.