Remove composite restriction
Summary:
The composite restriction is a result of a logical field and the fact that previous versions of SF supported primary keys with multiple columns. Since this package focuses on
on the drivers for the DBMSs which do not support these kinds of behaviors and we're deprecating the use of composites overall, these can safely go.

Test Plan: The files are just gone

Reviewers: cesar

Reviewed By: cesar

Differential Revision: https://phabricator.magic3w.com/D871