Removes the generic SQL components
Summary: While a nice idea, having a generic SQL underlying the mysql driver leqd to having to maintain an inflated codebase without any current benefits. Breaking it apart in the event of adding a PGSQL driver or so should be simple enough.

Test Plan: Let them rest, they're so tired!

Reviewers: cesar

Reviewed By: cesar

Differential Revision: https://phabricator.magic3w.com/D990