Exception handler prints the message
This uses the actual PHP internal serializer, which seems very similar
to the code we were planning to use. See my notes below.

---

It also includes the line that actually threw the exception. While this
is generally not too interesting, since the exception should be thrown
from a known good piece of code to to a bad call, but sometimes it can
help diagnosing what caused the issue.

For example, when a function like div(3, 0) throws a divide-by-null
exception, we do not thing that the issue is inside the div function, so
where inside it the issue was raised is actually not that interesting. But,
again, sometimes it can be helpful.