Resolves more phpstan issues
Spitfire is almost at the point where phpstan will just run without any
issue and we can raise the sensitivity of the reporting on it.

I also removed the belognstomany relationship. Currently it's a higher
priority to get something stable running and I'll be glad to add it back
in whenever the code is stable enough to support it.