Model relationships (#29)
* Renamed to active record

* The model factory needs to use active record too

* Added querybuilderinterface, added belongstoone::resolve

* BelongsToOne::resolveAll implemented

* Properly implemented BelongsToOneInjector

This change allows Spitfire applications to properly generate
restrictions for connected models. The relationship information is used
to generate the context for the subquery.

* Removed the superflous makeJoin method.

While I'm slightly emotionally attached to joins, the subquery provides
the superior isolation and is generally better at being, a subquery.

* Added a test for the absence method of belongstoone

I am not completely satisfied here, since the methods only differ by
three characters and are relatively long and cumbersome to test, making
it probable even that a mistake will make them differ and behave
inconsistently.

* Fixed unit tests

All the unit tests are back to working, which means that we're free to
make progress on new ventures and hopefully get those hasMany relations
working.

* Added missing resolve and resolveAll to HasMany

This should make it easy for us to query these relationships too.

* Model relationships finally work!

All the tests are passing, and the system now fully supports basic
belongs to one and has many relationships that can be reliably used to
query the database.

* Removed bad methods

The whereHas and where methods of relationship injectors where not being
used and causing us trouble.

* Remove deprecated functions

Scrutinizer pointed out severasl leftover issues that we can address to
get the code even leaner.