Refactoring request (#30)
* Extracted method and version to traits

This should allow our application to maintain all the related functions
in lean and easy to read components that the request method just brings
together.

This should make maintaining and debugging the small components
significantly easier.

It separates a request into being the sum of it's pieces. The request
now just holds a constructor, a fromGlobals method and a series of
traits

* Cleaned the code

We should be importing namespaces with more than two classes used as a
namespace and not as individual classes.

* Added assertions

This should make it easier for static analizers to make sure that the
code is sane and make it harder for me to screw the code up in hte
future by making it a bit stiffer.