Fuzzy counting. (#31)
* Queries can have tables and queries as source

We could use a union type, but introducing a dedicated union type feels
like a much cleaner implementation for what we need.

* Query Builder now quick counts

Quick counting allows to optimize queries by just having it
search until it found n records. For example, when a user has more
than 100 notifications, we actually no longer care if they have
100 or a 37,543.

It feels counter intutive, but this should actually increase
performance significantly when counting large sets.

* Assertions

We want to make some sanity checks on the application, ensuring that the
result is actually returned from the DBMS and that the primary key
actually exists.