Writing a file should issue a stream
When writing a stream to disk, the application has to detach the
original stream from the underlying handle, which makes it unusable.

The filesystem should return a new stream that the application can use.
Please note that this is a new stream on the SAME handle we were
originally using. Not on the file written to.