Fixes session read on first request
Reading the session during the first request would cause unexpected
behavior, because writing to the session and subsequently reading would
yield null.

This was relate to the fact that the cookie would not be set by the
browser mid request (obviously) so the system would fail a check to do
so. The simple solution is to check whether the session has been started
or the cookie is set.