Framework can detect the post limit was exceeded
If the post limit was exceeded the web server will truncate or remove
the body to ensure that the server is not overloaded.

By detecting this, we can alert the user that the request was invalid
due to the fact it exceeded the upload limits.

This means that future iterations should set their upload max filesize
to the maximum since we can otherrwise cannot detect the fact that a
upload was cut due to exceeeding the filesize